Extract logic for spotify OAuth flow
When a user makes a request that requires authentication for spotify, we
lead them through the OAuth flow. We create a link to spotify's auth
endpoint with the appropriate scopes (in our case to modify playlists).
We also start an HTTP server that receives the token from spotify on
authentication. We shutdown the HTTP server at the end of the flow so
that we don't have a hanging entryway.