Refactor data structure for enrolled users
Use an array as the underlying data structure.

I had started with an array, but found 2 key issues: the array pointer
syntax and O(n) removal of users. This led me to use a map instead. But
maps do not have an order and so I could not queue users the way I had
imagined.