Serve http auth server for lifetime of app
A user could exit the authentication flow prematurely. This left the
auth server running indefinitely. I decided to leave the server running
for the lifetime of the app instead of attempting to manage its state
manually. ie I opted for a simpler solution at the cost of having the
auth server always running.

While testing these changes, I also discovered that it is possible for a
user to navigate/GET the endpoint directly which would block the
goroutine responding to the request. I account for this with a
select/default block that would avoid writing to a channel that is not
being read from.