Create DiscordMessager for CommandContext
This change in parameter list makes the dependency of CommandContext on
a Messager more clear. It also affords us the freedom to get rid of the
NewContext function.

Another potential approach was to make DiscordMessager a field of the
DiscordBuilder and update the ChannelID per request. I decided against
this approach because the CommandContext is small in terms of space and
simple to build. Maybe the only other reason to adopt this approach is
to make a relationship more clear.