Remove redundant DiscordBuilder
As an abstraction, it was unclear what DiscordBuilder was responsible
for. The key responsibility it owned was the discordgo.Session.

Move to using discordgo directly in main.

Another side effect of this was the need to make the dependencies
globally scoped in order to access from the body of the handleMessage
function. I wonder if there is a stronger way to clear that up.