Add helpMsg field to command
Adding helpMsg led to a few other changes.

For one we could no longer run the command directly. The solution we
chose here was to create a method for the command type.

The other thing that had to change was how we registered commands. Our
approach here was to keep the command type hidden, pass the raw function
& helpMsg, then compose them up into the command held by the registry.